Refactored OperationMessageOptions into OperationMessage
